Added support for publishing NPM package.
