+++ HtmlMesh/index.html
@@ -3,7 +3,7 @@
 <head>
     <title>HTML Mesh Demo</title>
     <link rel="stylesheet" href="index.css">
-    <script type="module" src="index.js"></script>
+    <script type="module" src="example.js"></script>
 </head>
 <body>
     <canvas></canvas>
