+++ HtmlMesh/package.json
@@ -1,16 +1,37 @@
 {
-  "name": "htmlmesh",
-  "version": "1.0.0",
+  "name": "babylon-htmlmesh",
+  "version": "1.0.11",
   "description": "A BabylonJS Mesh that can display HTML DOM Content in scene (can occlude other meshes and can be occluded)",
   "type": "module",
+  "main": "dist/umd/babylon-htmlmesh.js",
+  "module": "dist/es/babylon-htmlmesh.js",
+  "exports": {
+    ".": {
+      "import": "./dist/es/babylon-htmlmesh.js",
+      "require": "./dist/umd/babylon-htmlmesh.js"
+    }
+  },
+  "files": [
+    "dist/es/babylon-htmlmesh.js",
+    "dist/umd/babylon-htmlmesh.js"
+  ],
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
     "preview": "vite preview",
-    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js"
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
+    "prepublishOnly": "mv README.md README-github.md && mv README-npm.md README.md",
+    "postpublish": "mv README.md README-npm.md && mv README-github.md README.md"
   },
   "author": "Eric Wood",
   "license": "ISC",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/ericwood73/BabylonJS-Extensions"
+  },
+  "peerDependencies": {
+    "@babylonjs/core": "^6.30.0"
+  },
   "devDependencies": {
     "@babylonjs/core": "^6.30.0",
     "@babylonjs/inspector": "^6.33.1",
@@ -21,9 +42,6 @@
     "ts-loader": "^9.5.1",
     "ts-node": "^10.9.1",
     "typescript": "^5.2.2",
-    "vite": "^5.0.0",
-    "webpack": "^5.89.0",
-    "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^4.15.1"
+    "vite": "^5.0.0"
   }
 }
