+++ HtmlMesh/tsconfig.json
@@ -19,5 +19,5 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["../src"]
+  "include": ["./src"]
 }
